fix: only send user icon related meta data
