+++ packages/nocodb-sdk/src/lib/index.ts
@@ -41,3 +41,4 @@ export * from '~/lib/errorUtils';
 export * from '~/lib/formBuilder';
 export * from '~/lib/ai';
 export * from '~/lib/import-export-data';
+export * from '~/lib/is';
